+++ src/test/scala/edu/rit/cs/dataunitylab/jsonoid/discovery/schemas/ArraySchemaSpec.scala
@@ -226,7 +226,8 @@ class ArraySchemaSpec extends UnitSpec with ScalaCheckPropertyChecks {
   }
 
   it should "convert tuple schemas using prefixItems" in {
-    (tupleSchema.toJson() \ "prefixItems")
+    (tupleSchema.toJson() \ "prefixItems") shouldNot equal(JNothing)
+    (tupleSchema.toJson() \ "items") shouldEqual(JBool(false))
   }
 
   it should "keep the count when transforming" in {