+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Switch to draft 2020-12
 - Switch back to `prefixItems`
 - Throw error when parsing with `unevaluatedProperties`
+- Set `items` to `false` in tuple schemas
 
 ## [0.16.0]
 ### Added